Backport 135f64e603266e0bb8852edf77d21a9cd79388ce

+++ src/hotspot/share/interpreter/bytecodes.cpp
@@ -99,12 +99,13 @@ int Bytecodes::special_length_at(Bytecodes::Code code, address bcp, address end)
       if (end != NULL && aligned_bcp + 3*jintSize >= end) {
         return -1; // don't read past end of code buffer
       }
-      jlong lo = (jint)Bytes::get_Java_u4(aligned_bcp + 1*jintSize);
-      jlong hi = (jint)Bytes::get_Java_u4(aligned_bcp + 2*jintSize);
-      jlong len = (aligned_bcp - bcp) + (3 + hi - lo + 1)*jintSize;
+      // Promote calculation to 64 bits to do range checks, used by the verifier.
+      int64_t lo = (int)Bytes::get_Java_u4(aligned_bcp + 1*jintSize);
+      int64_t hi = (int)Bytes::get_Java_u4(aligned_bcp + 2*jintSize);
+      int64_t len = (aligned_bcp - bcp) + (3 + hi - lo + 1)*jintSize;
       // only return len if it can be represented as a positive int;
       // return -1 otherwise
-      return (len > 0 && len == (int)len) ? len : -1;
+      return (len > 0 && len == (int)len) ? (int)len : -1;
     }
 
   case _lookupswitch:      // fall through
@@ -114,11 +115,12 @@ int Bytecodes::special_length_at(Bytecodes::Code code, address bcp, address end)
       if (end != NULL && aligned_bcp + 2*jintSize >= end) {
         return -1; // don't read past end of code buffer
       }
-      jlong npairs = (jint)Bytes::get_Java_u4(aligned_bcp + jintSize);
-      jlong len = (aligned_bcp - bcp) + (2 + 2*npairs)*jintSize;
+      // Promote calculation to 64 bits to do range checks, used by the verifier.
+      int64_t npairs = (int)Bytes::get_Java_u4(aligned_bcp + jintSize);
+      int64_t len = (aligned_bcp - bcp) + (2 + 2*npairs)*jintSize;
       // only return len if it can be represented as a positive int;
       // return -1 otherwise
-      return (len > 0 && len == (int)len) ? len : -1;
+      return (len > 0 && len == (int)len) ? (int)len : -1;
     }
   default:
     // Note: Length functions must return <=0 for invalid bytecodes.
